requirements.txt
psycopg2-binary
lxml